Move internal/containerfs to daemon/internal/containerfs
