+++ daemon/graphdriver/btrfs/btrfs.go
@@ -36,8 +36,8 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/graphdriver"
+	"github.com/docker/docker/daemon/internal/containerfs"
 	"github.com/docker/docker/daemon/internal/fstype"
-	"github.com/docker/docker/internal/containerfs"
 	"github.com/docker/go-units"
 	"github.com/moby/sys/mount"
 	"github.com/moby/sys/user"
