+++ daemon/graphdriver/vfs/driver.go
@@ -6,9 +6,9 @@ import (
 	"path/filepath"
 
 	"github.com/docker/docker/daemon/graphdriver"
+	"github.com/docker/docker/daemon/internal/containerfs"
 	"github.com/docker/docker/daemon/internal/quota"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/internal/containerfs"
 	"github.com/docker/go-units"
 	"github.com/moby/sys/user"
 	"github.com/opencontainers/selinux/go-selinux/label"
