+++ daemon/internal/containerfs/rm.go
