+++ daemon/pkg/plugin/backend_linux.go
@@ -21,10 +21,10 @@ import (
 	"github.com/containerd/log"
 	"github.com/containerd/platforms"
 	"github.com/distribution/reference"
+	"github.com/docker/docker/daemon/internal/containerfs"
 	v2 "github.com/docker/docker/daemon/pkg/plugin/v2"
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/internal/containerfs"
 	"github.com/docker/docker/pkg/authorization"
 	"github.com/docker/docker/pkg/pools"
 	"github.com/docker/docker/pkg/progress"
