+++ daemon/pkg/plugin/manager.go
@@ -15,10 +15,10 @@ import (
 	"github.com/containerd/containerd/v2/core/content"
 	"github.com/containerd/containerd/v2/plugins/content/local"
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/internal/containerfs"
 	"github.com/docker/docker/daemon/internal/lazyregexp"
 	v2 "github.com/docker/docker/daemon/pkg/plugin/v2"
 	"github.com/docker/docker/daemon/pkg/registry"
-	"github.com/docker/docker/internal/containerfs"
 	"github.com/docker/docker/pkg/authorization"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/api/types/events"
