+++ daemon/pkg/plugin/manager_linux_test.go
@@ -8,8 +8,8 @@ import (
 	"syscall"
 	"testing"
 
+	"github.com/docker/docker/daemon/internal/containerfs"
 	v2 "github.com/docker/docker/daemon/pkg/plugin/v2"
-	"github.com/docker/docker/internal/containerfs"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/api/types/backend"
